right_the_ship/requirements.txt
Django==5.0.7
annotated-types==0.7.0
asgiref==3.8.1
black==24.4.2
click==8.1.7
django-ninja==1.2.1
exceptiongroup==1.2.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pydantic==2.8.2
pytest==8.2.2
sqlparse==0.5.0
tomli==2.0.1
typing-extensions==4.12.2